wip: rename rule
Also tightened up matching based on feedback
