+++ tests/fixtures/rules/no-extra-reactive-curlies/invalid/test01-errors.json
