+++ tests/fixtures/rules/no-extra-reactive-curlies/invalid/test01-input.svelte
