+++ tests/fixtures/rules/no-extra-reactive-curlies/invalid/test02-errors.json
@@ -1,0 +1,7 @@
+[
+  {
+    "message": "Do not wrap reactive statements in curly braces unless necessary.",
+    "line": 3,
+    "column": 6
+  }
+]
