+++ tests/fixtures/rules/no-extra-reactive-curlies/invalid/test02-output.svelte
@@ -1,0 +1,4 @@
+<!-- prettier-ignore -->
+<script>
+  $: foo = info;
+</script>
