+++ tests/fixtures/rules/no-extra-reactive-curlies/valid/test01-input.svelte
@@ -1,0 +1,5 @@
+<!-- prettier-ignore -->
+<script>
+  $: foo1 = "bar"
+  $: foo2 = "bar";
+</script>
