+++ tests/fixtures/rules/no-extra-reactive-curlies/valid/test02-input.svelte
