+++ tests/fixtures/rules/no-extra-reactive-curlies/valid/test03-input.svelte
@@ -1,0 +1,12 @@
+<script>
+  $: {
+    let foo = "bar"
+  }
+
+  $: {
+    function fn() {}
+  }
+  $: {
+    class Klass {}
+  }
+</script>
