--- tests/fixtures/rules/no-unnecessary-reactive-curlies/valid/test01-input.svelte
@@ -1,3 +1,0 @@
-<script>
-  $: foo = info
-</script>
